Fix and rework hunger

+++ server/src/server/define.rs
@@ -26,7 +26,7 @@ pub static CARDINAL_DICT: phf::Map<char, u8> = phf_map! {
     'O' => 4,
 };
 
-pub const INCANTATION_REQ: [[u32; 7]; 7] = [
+pub const INCANTATION_REQ: [[u128; 7]; 7] = [
     [1, 0, 0, 0, 0, 0, 0],
     [2, 1, 1, 1, 0, 0, 0],
     [2, 2, 0, 1, 0, 2, 0],
@@ -207,7 +207,7 @@ pub const SEE_TAB_E: [(i32, i32); 81] = [
     (8, 8),
 ];
 // INVANTORY CONST
-pub const FOOD_INV: usize = 0; // symbolic, it is the hunger bar;
+pub const FOOD_INV: usize = 0;
 pub const T1_MAT_INV: usize = 1;
 pub const T2_MAT_INV: usize = 2;
 pub const T3_MAT_INV: usize = 3;
